New cyl domain

+++ VirtualWindTunnel/Components/CompCellSize.cs
@@ -67,7 +67,7 @@ namespace Eddy
 
             double desiredCellSize = 1;
             double blockMeshCellSize = 1;
-            double acc = 1;
+            int acc = 1;
 
             OFCylDomain CylDom;
             OFBoxDomain BoxDom;
@@ -101,7 +101,7 @@ namespace Eddy
 
             
 
-            acc = Math.Round(((Math.Log(blockMeshCellSize) - Math.Log(desiredCellSize)) / Math.Log(2)));
+            acc =(int)(Math.Round(((Math.Log(blockMeshCellSize) - Math.Log(desiredCellSize)) / Math.Log(2))));
